Syntax: Support interpreting mxcells that have been made into links
In draw.io, you can right click on a node and hit "Edit Link" to have a
node pop up with a link to another diagram page when focused. This is
extremely nice for components that represent containers on another page
of the diagram.

When you make a node into a link, it changes the layout of the XML a
little bit. The mxCell nodes that we usually read have their label & ID
lifted to a wrapping UserObject node.

    <UserObject label="..." id="...">
      <mxCell style="..."></mxCell>
    </UserObject>

So, the current implementation would panic upon encountering it.

This changes the parser to also look at these UserObject nodes, and pass
them as a parent node to `cell_from_elem`, so such nodes should be
properly recognized now.